requirements.txt
chainlit
python-dotenv
requests
SQLAlchemy
asyncpg
transformers
torch
accelerate
sentencepiece
protobuf
